refactor hotkeys setting
